move: uiModel 패키지 생성
