+++ app/build.gradle.kts
@@ -70,4 +70,6 @@ dependencies {
     androidTestRuntimeOnly("de.mannodermaus.junit5:android-test-runner:1.3.0")
     testImplementation("androidx.arch.core:core-testing:2.2.0")
     implementation("com.squareup.leakcanary:leakcanary-android:2.14")
+    implementation("androidx.room:room-runtime:2.6.1")
+    kapt("androidx.room:room-compiler:2.6.1")
 }
