+++ app/src/main/java/woowacourse/shopping/productdetail/ProductDetailMapper.kt
@@ -1,5 +1,6 @@
 package woowacourse.shopping.productdetail
 
 import woowacourse.shopping.domain.ShoppingCartItem
+import woowacourse.shopping.productdetail.uimodel.CountResultUiModel
 
 fun ShoppingCartItem.toCountResult(): CountResultUiModel = CountResultUiModel(this.totalQuantity, this.product.price)
