+++ app/src/main/java/woowacourse/shopping/productdetail/uimodel/RecentProductState.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.productdetail
+package woowacourse.shopping.productdetail.uimodel
 
 import woowacourse.shopping.util.UiState
