+++ app/src/main/java/woowacourse/shopping/productlist/ProductListMapper.kt
@@ -3,6 +3,8 @@ package woowacourse.shopping.productlist
 import woowacourse.shopping.domain.Product
 import woowacourse.shopping.domain.RecentProduct
 import woowacourse.shopping.domain.ShoppingCartItem
+import woowacourse.shopping.productlist.uimodel.ProductUiModel
+import woowacourse.shopping.productlist.uimodel.RecentProductUiModel
 
 fun Product.toProductUiModel(cartItemCount: Int): ProductUiModel =
     ProductUiModel(
