+++ app/src/main/java/woowacourse/shopping/productlist/uimodel/LoadProductState.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.productlist
+package woowacourse.shopping.productlist.uimodel
 
 import woowacourse.shopping.util.UiState
