+++ app/src/main/java/woowacourse/shopping/productlist/uimodel/ProductUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.productlist
+package woowacourse.shopping.productlist.uimodel
 
 data class ProductUiModel(
     val id: Long,
