+++ app/src/main/java/woowacourse/shopping/shoppingcart/uimodel/LoadCartItemState.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.shoppingcart
+package woowacourse.shopping.shoppingcart.uimodel
 
 import woowacourse.shopping.util.UiState
