+++ app/src/test/java/woowacourse/shopping/viewmodel/ProductListViewModelTest.kt
@@ -6,9 +6,9 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import woowacourse.shopping.DummyProductRepository
 import woowacourse.shopping.DummyShoppingRepository
-import woowacourse.shopping.productlist.LoadProductState
 import woowacourse.shopping.productlist.ProductListViewModel
 import woowacourse.shopping.productlist.toProductUiModel
+import woowacourse.shopping.productlist.uimodel.LoadProductState
 import woowacourse.shopping.viewmodel.fixtures.InstantTaskExecutorExtension
 import woowacourse.shopping.viewmodel.fixtures.getOrAwaitValue
